# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "",
  "basePath": "",
  "tags": [
    {
      "name": "",
      "description": ""
    }
  ],
  "schemes": [],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-API-KEY",
      "description": "API KEY"
    }
  },
  "paths": {
    "/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/save-url/{url}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "url",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "URL to download"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/status/{job_id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The job id to return a status for!"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    }
  }
}